Tests for natural sorting
